fixed no return to origin bug
The cleaningProtocol Method was moved from Robomemory class to SweepMonter class, due to the need of calling the goBack() method to return to origin.

A problem with the goBack function  is that either we need a bigger floorplan or different floorplans for testing. This also applies to the isblocked() function. Both functions are in the SweepMonster Class.